MBP-230 Better error messages when a user doesn't have delete permission
git-svn-id: svn+ssh://svn.biomatters.com/usr/local/svnroot/geneious/trunk/plugins/MooreaLabBench@43196 84f1a7b3-ea07-0410-887d-f1e6e76816dd

+++ src/com/biomatters/plugins/biocode/assembler/lims/AddAssemblyResultsToLimsOperation.java
@@ -177,6 +177,9 @@ public class AddAssemblyResultsToLimsOperation extends DocumentOperation {
         if(assemblyResult.workflowId == null || assemblyResult.extractionId == null) {
             return 0;
         }
+        if(!BiocodeService.getInstance().deleteAllowed("assembly")) {
+            throw new DocumentOperationException("It appears that you do not have permission to delete sequence records.  Please contact your System Administrator for assistance");
+        }
 
         String sql = "DELETE FROM assembly WHERE workflow=? AND extraction_id=?";
         LIMSConnection limsConnection = BiocodeService.getInstance().getActiveLIMSConnection();
